everything is imported and ready to start push 2
I have imported everything across into this new repo, and the game can be worked on further from this point.

NOTE: there are 2 commits that I have done on my home computer that cannot be pushed up to the cloud. they may be pushed later on, or not at all, I don't know how that will work, or even if this will push before the other 2. we will see.