Weapon switcher reduced, Created a utility switcher script
I created a utility switcher script where the player can now switch in between the torch and the NV camera. The player can switch between the different utilities by the number bar, 1,2,3, etc. In order to do this I had to remove the number bar weapon switch method for the weapon script. so know the weapons can only be changed by the scroll wheel.